msgviewer: show error if MessageInfo has error
MessageInfo structs can be returned from the workers with parser errors.
MessageViews were previously prevented from opening if the error field
was not nil, however the addition of message view splits have allowed
these messages to have split views created. A panic can occur if a split
is open and one of these messages is scrolled over.

Prevent access to MessageInfo methods if the Error field is set.
Display the error to the user in the split view. The partSwitcher will
be nil if an error is set: check for this condition before calling it's
methods.

Reported-by: Ben Lee-Cohen <ben@lee-cohen.com>
Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>
Acked-by: Robin Jarry <robin@jarry.cc>