maildir: keep less data in memory for sorting
Sorting opens and reads portions of every file within a directory in
order to gather the data needed. Specifically, RFC822Headers and
BodyStructure are not needed. The RFC822Headers field stores a lot of
information, and the BodyStructure field requires parsing the body of
the email.

Don't set these two values when parsing.

Note: in my testing, this dropped sorting a 52k archive from 2.2gb of
ram usage, to < 500mb

Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>
Acked-by: Robin Jarry <robin@jarry.cc>