search: handle date ranges in search/filter query
Handle date ranges in the filter and search commands for searching and
filtering based on the Date: header. Implement a flag (-d) that accepts
a date range <start[..end]> where the start date is included in the
range but the end date is not, i.e. [start,end).

The start or end date can be omitted: "start", "start..", "..end", or
"start..end" are all valid inputs.

An example filter query would look like this: :filter -d 2022-11-09..

The dates should be in the YYYY-MM-DD format.

Signed-off-by: Koni Marti <koni.marti@gmail.com>
Acked-by: Robin Jarry <robin@jarry.cc>