attach: open file picker menu with -m flag
Open a user-defined file picker with the -m flag for the attach command
to select attachments.

Specify your file picker of choice with the 'file-picker-cmd' in the
[composer] section of aerc.conf, e.g.  "file-picker-cmd=fzf -m".

A '%s' placeholder can be used in the 'file-picker-cmd' which is then
substituted for the argument <arg> provided to :attach -m <arg>.

For example, when you set 'file-picker-cmd=find %s -type f | fzf -m',
you can easily change the directory to start the search with ':attach -m
<path-to-search>'.

Tested with fzf, fzy and peco.

Implements: https://todo.sr.ht/~rjarry/aerc/108
Signed-off-by: Koni Marti <koni.marti@gmail.com>
Acked-by: Tim Culverhouse <tim@timculverhouse.com>