config: fix [filters] ordering
Using KeysHash() does not guarantee any stable ordering of elements.
This causes random weirdness when some filters are using wildcards:

  [filters]
  text/plain=colorize
  text/html=html
  text/*=bat -fP --file-name="$AERC_FILENAME" --style=plain

When the source order is not preserved, the text/* filter may be matched
first against text/plain or text/html parts.

Use Keys() which does not create a map and preserve original ordering.

Fixes: 17bb9387c4a3 ("config: move [filters] parsing to separate file")
Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Bence Ferdinandy <bence@ferdinandy.com>