colorize: add solarized theme
This modifies the colorize script to accept a command line parameter to
change the colour theme. Currently only a solarized version is added.
Due to the nature of awk the theme has to be defined via the `-v` flag.

Due to the `switch` statement only being available in GNU awk we use a
`if else` statement to ensure that the default colours are used if
either the `THEME` variable is not set at all or set to `default`.

Solarized colour scheme: https://ethanschoonover.com/solarized/

Signed-off-by: Jens Grassel <jens@wegtam.com>
Signed-off-by: Robin Jarry <robin@jarry.cc>