msgstore: set store.uids in threaded directories
Commit b46e57324394 ("store: fix server-side threads toggling")
introduced a regression where the uids of the store were not set after
rebuilding the threads. This would result in a flash of pending messages
as they were repopulated. This also would not properly remove messages
which were moved out of the store (deleted, moved, archived) with server
side threading.

Re-add the line that sets the store.uids.

Fixes: b46e57324394 ("store: fix server-side threads toggling")
Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>
Acked-by: Bence Ferdinandy <bence@ferdinandy.com>