split: refactor split update logic
Refactor split update logic to more simply update the split. Through the
evolution of the split logic, additional variables were stored within
the account which allows for cleaner updating of the split.

Compare selected UID instead of pointer to message when deciding not to
update split.

Allow splits to be created and closed when no message is selected. The
split will be filled with a ui.Fill (blank). The user will only see a
border at the split location when no message is selected.

Rename clearSplit to closeSplit, as it is only used in the case when the
user doesn't want a split anymore.

Ensure that the selected UID is reset to the magic UID when there are no
messages left in the message store.

Signed-off-by: Tim Culverhouse <tim@timculverhouse.com>
Tested-by: Bence Ferdinandy <bence@ferdinandy.com>